refactor(agent): extract 44 write action handlers into registry-based families (PR 5)
Continues the strangler fig decomposition of agentExecutor.ts (4485 → ~2520 lines).
Extracts all write/mutation cases from the inline switch into five handler files —
taskWriteActions, projectWriteActions, goalAreaActions, captureActions, and
agentControlActions — each registered via ActionRegistry.registerRaw(). Introduces
agentTypes.ts for shared types (breaks circular deps), a WriteActionExecutor shim
that exposes private executor methods to handlers via runtime.exec, and extends
ActionRuntime with captureService, projectService, persistencePrisma, and exec.
Moves triageCaptureText/suggestCaptureRoute utilities and IDEMPOTENT_PLANNER_APPLY_ACTIONS
out of the executor file scope. Removes two now-dead private methods (handleCreateTask,
handleCreateProject).
